Add clean_and_process_counts_for_dge function
Function converts a df containing counts values to numerics (except for the first two cols). Provides a summary of the data before and after cleaning. Counts the number of zeros in each column and checks for problematic rows where sgRNA or Gene values are zero. Removes rows where sgRNA or Gene is zero and replaces zero counts with a small value in the other columns. Converts the cleaned DataFrame into a matrix, excluding sgRNA and Gene columns and setting the Gene column as rownames.