Change absolute file paths
Created and configured a yml file that contains the absolute file paths on the NIH laptop. This avoids putting the absolute paths in individual R files. Instead data can be imported with the same few lines even if the paths change. For future reference, only modify the path in the stmn2_config.yml file